+++ main.py
@@ -1,5 +1,5 @@
-from gui import launch_app
-from app_config import load_config
+from gui.window import launch_app
+from config.app_config import load_config
 
 if __name__ == "__main__":
     load_config()
